--- public/docs/_examples/reactive-forms/ts/app/hero-list.component.1.html
@@ -1,8 +1,0 @@
-<!-- #docregion -->
-<nav>
-  <a *ngFor="let hero of heroes | async" (click)="select(hero)">{{hero.name}}</a>
-</nav>
-
-<div *ngIf="selectedHero">
-  <hero-detail [hero]="selectedHero"></hero-detail>
-</div>
