--- public/docs/_examples/reactive-forms/ts/app/hero-list.component.html
@@ -1,17 +1,0 @@
-<!-- #docregion -->
-<h3 *ngIf="isLoading"><i>Loading heroes ... </i></h3>
-<h3 *ngIf="!isLoading">Select a hero:</h3>
-
-<nav>
-  <button (click)="getHeroes()" class="btn btn-primary">Refresh</button>
-  <a *ngFor="let hero of heroes | async" (click)="select(hero)">{{hero.name}}</a>
-</nav>
-
-<div *ngIf="selectedHero">
-  <hr>
-  <h2>Hero Detail</h2>
-  <h3>Editing: {{selectedHero.name}}</h3>
-  <!-- #docregion hero-binding -->
-  <hero-detail [hero]="selectedHero"></hero-detail>
-  <!-- #enddocregion hero-binding -->
-</div>
