--- public/docs/_examples/reactive-forms/ts/app/main-final.ts
@@ -1,5 +1,0 @@
-// tslint:disable:no-unused-variable
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { AppModule } from './app.module';
-
-platformBrowserDynamic().bootstrapModule(AppModule);
