+++ public/docs/_examples/reactive-forms/ts/final.plnkr.json
@@ -1,5 +1,6 @@
 {
   "description": "Angular Reactive Forms (final)",
+  "basePath": "src/",
   "files":[
     "styles.css",
 
@@ -12,7 +13,7 @@
     "app/hero-list.component.html",
     "app/hero-list.component.ts",
 
-    "app/main-final.ts",
+    "main-final.ts",
     "index-final.html"
   ],
   "main": "index-final.html",
