+++ public/docs/_examples/reactive-forms/ts/src/app/hero-list.component.ts
@@ -1,0 +1,32 @@
+// #docregion
+import { Component, OnInit } from '@angular/core';
+import { Observable }        from 'rxjs/Observable';
+import 'rxjs/add/operator/finally';
+
+import { Hero }        from './data-model';
+import { HeroService } from './hero.service';
+
+@Component({
+  moduleId: module.id,
+  selector: 'hero-list',
+  templateUrl: 'hero-list.component.html'
+})
+export class HeroListComponent implements OnInit {
+  heroes: Observable<Hero[]>;
+  isLoading = false;
+  selectedHero: Hero;
+
+  constructor(private heroService: HeroService) { }
+
+  ngOnInit() { this.getHeroes(); }
+
+  getHeroes() {
+    this.isLoading = true;
+    this.heroes = this.heroService.getHeroes()
+                      // Todo: error handling
+                      .finally(() => this.isLoading = false);
+    this.selectedHero = undefined;
+  }
+
+  select(hero: Hero) { this.selectedHero = hero; }
+}
