+++ public/docs/_examples/reactive-forms/ts/src/app/main-final.ts
@@ -1,0 +1,5 @@
+// tslint:disable:no-unused-variable
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { AppModule } from './app.module';
+
+platformBrowserDynamic().bootstrapModule(AppModule);
