+++ public/docs/_examples/reactive-forms/ts/src/index.html
@@ -20,7 +20,7 @@
 
     <script src="systemjs.config.js"></script>
     <script>
-      System.import('app').catch(function(err){ console.error(err); });
+      System.import('main.js').catch(function(err){ console.error(err); });
     </script>
   </head>
